app/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
idna==2.10
Markdown==3.3.4
pycodestyle==2.7.0
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.5
